+++ ios/Classes/SwiftImageDownloaderPlugin.swift
@@ -1,4 +1,5 @@
 import Flutter
+import MobileCoreServices
 import Photos
 import UIKit
 
@@ -14,7 +15,7 @@ public class SwiftImageDownloaderPlugin: NSObject, FlutterPlugin {
         case "downloadImage":
 
             guard let dict = call.arguments as? [String: Any], let url = dict["url"] as? String else {
-                result(FlutterError(code: "assertion_error", message: "url required.", details: nil))
+                result(FlutterError(code: "assertion_error", message: "url is required.", details: nil))
                 return
             }
 
@@ -23,15 +24,46 @@ public class SwiftImageDownloaderPlugin: NSObject, FlutterPlugin {
                 case .authorized:
                     self.downloadImage(url, result)
                 case .denied, .restricted:
-                    result(false)
+                    result(nil)
                 case .notDetermined:
-                    result(false)
+                    result(nil)
+                }
+            }
+        case "findPath":
+            setData(call, result: result) { imageId in
+                findPath(imageId: imageId) { value in
+                    result(value)
+                }
+            }
+        case "findName":
+            setData(call, result: result) { imageId in
+                let name = findName(imageId: imageId)
+                result(name)
+            }
+        case "findByteSize":
+            setData(call, result: result) { imageId in
+                findByteSize(imageId: imageId) { value in
+                    result(value)
+                }
+            }
+        case "findMimeType":
+            setData(call, result: result) { imageId in
+                findMimeType(imageId: imageId) { value in
+                    result(value)
                 }
             }
         default:
             result(FlutterMethodNotImplemented)
             break
         }
+    }
+
+    private func setData(_ call: FlutterMethodCall, result: @escaping FlutterResult, completion: (String) -> Void) {
+        guard let imageId = (call.arguments as? Dictionary<String, String>)?["imageId"] else {
+            result(FlutterError(code: "assertion_error", message: "imageId is required.", details: nil))
+            return
+        }
+        completion(imageId)
     }
 
     private func downloadImage(_ url: String, _ result: @escaping FlutterResult) {
@@ -94,6 +126,7 @@ public class SwiftImageDownloaderPlugin: NSObject, FlutterPlugin {
             }
         }
 
+        var imageId: String?
         PHPhotoLibrary.shared().performChanges({
             let request = PHAssetChangeRequest.creationRequestForAsset(from: image)
             if let exifDatetime = exifDatetime {
@@ -105,9 +138,11 @@ public class SwiftImageDownloaderPlugin: NSObject, FlutterPlugin {
             if let latitude = latitude, let logitude = logitude {
                 request.location = CLLocation(latitude: CLLocationDegrees(latitude), longitude: CLLocationDegrees(logitude))
             }
+            imageId = request.placeholderForCreatedAsset?.localIdentifier
+
         }) { isSuccess, error in
-            if isSuccess {
-                result(true)
+            if let imageId = imageId, isSuccess {
+                result(imageId)
             } else {
                 result(FlutterError(code: "save_error", message: "Couldn't save to photo library.", details: error))
             }
@@ -130,18 +165,47 @@ public class SwiftImageDownloaderPlugin: NSObject, FlutterPlugin {
             return
         }
         do {
+            var imageId: String?
             try data.write(to: imageUrl)
             PHPhotoLibrary.shared().performChanges({
-                PHAssetChangeRequest.creationRequestForAssetFromImage(atFileURL: imageUrl)
-            }, completionHandler: { _, error in
-                if error != nil {
-                    result(FlutterError(code: "save_error", message: "Couldn't save gif to photo library.", details: error))
+                let request = PHAssetChangeRequest.creationRequestForAssetFromImage(atFileURL: imageUrl)
+                imageId = request?.placeholderForCreatedAsset?.localIdentifier
+            }, completionHandler: { isSuccess, error in
+                if let imageId = imageId, isSuccess {
+                    result(imageId)
                 } else {
-                    result(true)
+                    result(FlutterError(code: "save_error", message: "Couldn't save to photo library.", details: error))
                 }
             })
         } catch {
             result(FlutterError(code: "save_error", message: "Couldn't save gif to photo library.", details: error))
+        }
+    }
+
+    private func findPath(imageId: String, completion: @escaping (String) -> Void) {
+        let asset = PHAsset.fetchAssets(withLocalIdentifiers: [imageId], options: nil).firstObject!
+        asset.requestContentEditingInput(with: PHContentEditingInputRequestOptions()) { input, _ in
+            let url = input?.fullSizeImageURL
+            completion(url!.absoluteString)
+        }
+    }
+
+    private func findName(imageId: String) -> String {
+        return PHAsset.fetchAssets(withLocalIdentifiers: [imageId], options: nil).firstObject?.value(forKey: "filename") as! String
+    }
+
+    private func findByteSize(imageId: String, completion: @escaping (Int) -> Void) {
+        let asset = PHAsset.fetchAssets(withLocalIdentifiers: [imageId], options: nil).firstObject!
+        PHImageManager.default().requestImageData(for: asset, options: nil) { imageData, _, _, _ in
+            completion(imageData!.count)
+        }
+    }
+
+    private func findMimeType(imageId: String, completion: @escaping (String) -> Void) {
+        let asset = PHAsset.fetchAssets(withLocalIdentifiers: [imageId], options: nil).firstObject!
+        PHImageManager.default().requestImageData(for: asset, options: nil) { _, uti, _, _ in
+            let mimeType: String = UTTypeCopyPreferredTagWithClass(uti! as CFString, kUTTagClassMIMEType)!.takeRetainedValue() as String
+            completion(mimeType)
         }
     }
 }